+++ CheckPoint/urls.py
@@ -16,6 +16,8 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path, include
+from django.conf import settings
+from django.conf.urls.static import static
 
 urlpatterns = [
     path('admin/', admin.site.urls),
@@ -25,6 +27,11 @@ urlpatterns = [
     path('roms/', include('CheckPoint.roms.urls'), name='roms page'),
     path('saves/', include('CheckPoint.saves.urls'), name='saves page'),
 ]
+
+# this is so we can serve media files in development
+if settings.DEBUG:
+    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
+
 
 handler404 = 'CheckPoint.common.views.custom_404'